Allow to configure devspace checkout options and change default
As part of the work on the playbook refactoring, c829116 introduced options to
force the update of the devspace. A side effect is that any redeployment of the
devspace resets the CI space and makes the role effectively a provisioning playbook.

While we might want to clarify the separation between the provisioning and the
deployment steps in the future, this commit should restore the default
assumptions (i.e. redeploy only) and make the playbook idempotent by default.
Force redeploying can be configured explicitly via the new devspace options.