Ara: add plugin to default configuration
If ara is not installed, this should produce no change in behavior.

If ara *is* installed, then all calls to `ansible-playbook` will be
recorded to an sqlite database in ~/.ara/.

If additionally, ARA_DATABASE is configured to point at a shared
database, then all runs of ansible-playbook will be centrally recorded.