Prefix decoupled Ansible roles with openmicroscopy.
While trying to deploy the Sphinx changes on the ci-docs nodes, I encountered
execution issues while trying to execute the ci-deloyment playbook.

This commit prefixes all roles installed via Galaxy with openmicroscopy. It
also adds a tag to the Sphinx role as this was the only required change in that
case.